+++ RELEASENOTES.md
@@ -2,4 +2,6 @@
 
 * Fixed issue when full screen mode is activated which may cause black bar to appear on either size (top/bottom or left/right).
 
+* Fixed preferences pop out title bar color customization setting not saving.
+
 * Added preference option in pop out settings to disable pop out progress messages from appearing.